Move newly added methods on Point into ext trait
