+++ embedded-graphics-core/src/geometry/size.rs
@@ -172,7 +172,7 @@ impl Size {
     }
 
     /// Creates a size from two corner points of a bounding box.
-    pub const fn from_bounding_box(corner_1: Point, corner_2: Point) -> Self {
+    pub(crate) const fn from_bounding_box(corner_1: Point, corner_2: Point) -> Self {
         let width = (corner_1.x - corner_2.x).abs() as u32 + 1;
         let height = (corner_1.y - corner_2.y).abs() as u32 + 1;
